Add class declarations and delete unneccessary code
I added in the class declarations for ScreenNormal and ScreenAdmin.
ScreenAdmin is a child of ScreenNormal. I'm adding these in because it
seemed like it would be much better to put these as classes instead of a
bunch unconnected functions. The unneccessary code I deleted was
commented out code that I wasn'nt really using.